+++ src/vue/components/containment/card/index.ts
@@ -1,0 +1,6 @@
+export { default as GkCard } from './GkCard.vue'
+export { default as GkCardActions } from './GkCardActions.vue'
+export { default as GkCardItem } from './GkCardItem.vue'
+export { default as GkCardSubtitle } from './GkCardSubtitle.vue'
+export { default as GkCardText } from './GkCardText.vue'
+export { default as GkCardTitle } from './GkCardTitle.vue'